backend/requirements.txt
Flask==3.0.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Marshmallow==1.0.0
marshmallow-sqlalchemy==1.0.0
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.1.2
celery==5.3.6
redis==5.0.1
boto3==1.34.34
gunicorn==21.2.0
pytest==8.0.0
python-jose==3.3.0
email-validator==2.1.0.post1
Flask-Swagger-UI==4.11.1 